Add a run_dir attribute
The run_dir is a directory that will be created (if it doesn't already
exist) on start. If they aren't already absolute, we assume that
pid_file, stderr_file, and stdout_file go in the run_dir.

Typically, pidfiles go in /var/run or a subdirectory of it. But if your
system puts /var/run on a tmpfs, then it will be empty on boot. We
therefore create a directory where pidfiles etc can go so you don't
have to worry about it.

Fixes GH #14